+++ src/index.ts
@@ -39,8 +39,8 @@ import DatePicker from './Molecules/DatePicker';
 import Drawer from './Molecules/Drawer';
 import FeedbackBanner from './Molecules/FeedbackBanner';
 import Fieldset from './Molecules/Fieldset';
-import FormField from './Molecules/FormField';
 import FlexTable from './Molecules/FlexTable';
+import FormField from './Molecules/FormField';
 import InfoBar from './Molecules/InfoBar';
 import Input from './Molecules/Input';
 import LabeledValue from './Molecules/LabeledValue';
@@ -53,13 +53,14 @@ import Modal from './Molecules/Modal';
 import MultiStepProgress from './Molecules/MultiStepProgress';
 import Number from './Molecules/Number';
 import PageHeaderCard from './Molecules/PageHeaderCard';
+import PageWrapper from './Molecules/PageWrapper';
 import Pagination from './Molecules/Pagination';
 import ProgressBar from './Molecules/ProgressBar';
-import PageWrapper from './Molecules/PageWrapper';
 import Rating from './Molecules/Rating';
 import SegmentedControl from './Molecules/SegmentedControl';
 import Select from './Molecules/Select';
 import SelectionCard from './Molecules/SelectionCard';
+import ShowMoreButton from './Molecules/ShowMoreButton';
 import Slider from './Molecules/Slider';
 import Switch from './Molecules/Switch';
 import Tabs from './Molecules/Tabs';
@@ -86,6 +87,7 @@ import { numberWithLimit } from './common/utils';
 
 /** Runtime code */
 export {
+  /** @deprecated */ Select,
   Accordion,
   AccordionItem,
   Avatar,
@@ -97,64 +99,66 @@ export {
   CardWithTabs,
   CardWithTitle,
   CollapsibleCard,
-  createTheme,
   CssGrid,
-  DateTime,
   DatePicker,
+  DateTime,
+  Development,
   Drawer,
   DropdownBubble,
-  Development,
   FadedScroll,
   FeedbackBanner,
   Fieldset,
   Flag,
-  Flexbox,
   FlexTable,
+  Flexbox,
   FormField,
   FormLabel,
-  IsomorphicMedia,
   Icon,
   InfoBar,
   Input,
+  IsomorphicMedia,
   LabeledValue,
   Legend,
   LineScale,
   Link,
   LinkBuy,
+  LinkContext,
+  LinkProps,
+  LinkProvider,
+  LinkProviderProps,
   LinkSell,
   List,
   ListItem,
   ListWithTitles,
   Media,
-  Number,
-  MultiStepProgress,
   Modal,
+  MultiStepProgress,
+  Number,
   PageHeaderCard,
   PageWrapper,
   Pagination,
-  Portal,
   Pill,
+  Portal,
   ProgressBar,
   Rating,
   SegmentedControl,
-  /** @deprecated */ Select,
   SelectionCard,
   Separator,
+  ShowMoreButton,
   Skeleton,
   Slider,
   Spinner,
   StatusModal,
   Switch,
+  TabTitle,
   Table,
   Tabs,
   TabsNav,
-  TabTitle,
   Tbody,
   Td,
   Tfoot,
   Th,
   Thead,
-  theme,
   Time,
   Timeline,
   Tooltip,
@@ -163,15 +167,13 @@ export {
   Truncate,
   TruncateWithTooltip,
   Typography,
-  useKeyPress,
-  useMedia,
-  useIsomorphicMedia,
-  useOnClickOutside,
   VisuallyHidden,
-  LinkContext,
-  LinkProvider,
-  LinkProviderProps,
-  LinkProps,
-  useLink,
+  createTheme,
   numberWithLimit,
+  theme,
+  useIsomorphicMedia,
+  useKeyPress,
+  useLink,
+  useMedia,
+  useOnClickOutside,
 };